Add files via upload
Initializing variables to 0 are found in the DropAttrib() macro in the beginning. Then starting just before the 1999 data step, in a new macro call Attrib(). 2017 does not include any macros so they're in there by themselves and  I'm not sure why but for 2015 & 2016 I had to duplicate this info in their data step to get the missing values to change to 0:
        if distVeg eq . then distVeg=0;
         if distHT eq . then distHT=0;
         if distVeg2 eq . then distVeg2=0;
         if distHT2 eq . then distHT2=0;